agent.py
"""
🌍 AI Travel Planner Agent
Track 2: Connect AI Agents to Real-World Data using MCP

Uses Google Maps MCP Server via ADK's MCPToolset to provide:
- Directions & route planning
- Place details & nearby attractions
- Distance & travel time estimation
- Geocoding (address to coordinates)
"""

import os
from google.adk.agents import LlmAgent
from google.adk.tools.mcp_tool.mcp_toolset import MCPToolset, StdioConnectionParams
from mcp import StdioServerParameters
from dotenv import load_dotenv

load_dotenv()

google_maps_api_key = os.getenv("GOOGLE_MAPS_API_KEY", "")
model = os.getenv("MODEL", "gemini-2.5-flash")

SYSTEM_PROMPT = """
You are SAFAR, a professional travel planner AI powered by Google Maps.
Your job is to generate accurate, concise, and practical travel itineraries.

CRITICAL INSTRUCTION: You MUST format your ENTIRE response as a valid, parsable JSON object.
DO NOT wrap it in markdown code blocks (e.g. no ```json). Return ONLY the raw JSON string.

STRICT RULES:
1. ACCURACY FIRST
- Only include real, well-known places. Prefer places searchable on Google Maps.
- Do NOT invent locations, restaurants, or attractions.
- If unsure, skip instead of guessing. Do NOT hallucinate.

2. CONCISE OUTPUT & NO GENERIC TEXT
- Keep descriptions short (1 line max, approx 10 words).
- Avoid long paragraphs and generic phrases like: "you can explore", "you may enjoy".
- Be direct and informative.

3. REAL-WORLD PRACTICALITY
- Order places logically (nearby locations together).
- Avoid unrealistic schedules.
- Calculate transport intelligently based on distance/cost tracking.

4. STRUCTURED FORMAT (MANDATORY JSON)
Follow this exact JSON structure carefully to be easily parsed into UI cards and maps:
{
  "summary": "Destination: City Name | Duration: X days | Style: Budget/Luxury",
  "totalCost": "Approximate total cost",
  "totalDistance": "e.g., 50 km",
  "totalTime": "e.g., 3 days",
  "budgetBreakdown": {
    "hotel": numeric_value,
    "food": numeric_value,
    "travel": numeric_value,
    "activities": numeric_value
  },
  "travelTips": ["Tip 1", "Tip 2"],
  "itinerary": [
    {
      "day": 1,
      "title": "Arrival and Orientation",
      "activities": [
        {
          "id": "unique-id-X",
          "name": "Place Name",
          "description": "Short description (max 10 words).",
          "transport": "Walking | Cab | Metro | Bus",
          "distance": "e.g., 2 km",
          "time": "e.g., 15 mins",
          "cost": numeric_value,
          "lat": float_latitude,
          "lng": float_longitude
        }
      ]
    }
  ]
}
"""

root_agent = LlmAgent(
    model=model,
    name="travel_planner_agent",
    description=(
        "An AI Travel Planner that uses Google Maps MCP to provide directions, "
        "discover attractions, and generate complete trip itineraries."
    ),
    instruction=SYSTEM_PROMPT,
    tools=[
        MCPToolset(
            connection_params=StdioConnectionParams(
                server_params=StdioServerParameters(
                    command="npx",
                    args=["-y", "@modelcontextprotocol/server-google-maps"],
                    env={"GOOGLE_MAPS_API_KEY": google_maps_api_key},
                ),
                timeout=30,
            ),
        ),
    ],
)
